Implemented property provider for Kafka
Supplying properties via PropertyProvider works for KafkaProducer and
KafkaConsumer. 

In KafkaProducer, if connection to the server is lost, we will check to
see if there are new properties available. If there are, we will create
a new client. 

KafkaConsumer will do the same thing, but we need to update KafkaClients
JARs to pick up a bug fix where consumer.poll() hangs.